signature

+++ app/Http/Controllers/ProjectController.php
@@ -27,19 +27,26 @@ class ProjectController extends Controller
         $attributes = Kozijnen::all();
         $attributeValues = $request->except(['_token', 'KlantNaam', 'ProjectAdres', 'Email']);
 
-        foreach ($attributes as $attribute) {
-            $value = $attributeValues[$attribute->kozijn] ?? null;
-            $project->kozijnen()->attach($attribute->id, ['value' => $value]);
+        if ($attributeValues){
+            foreach ($attributes as $attribute) {
+                $value = $attributeValues[$attribute->kozijn] ?? null;
+                $project->kozijnen()->attach($attribute->id, ['value' => $value]);
+            }
         }
 
         return redirect('/')->with('message', 'Project created successfully!');
     }
     public function show(Project $project)
     {
-        $kozijnen = $project->kozijnen;
-        $akkoord = $project->akkoord;
-        $info = $project->info;
-        return view('projects/show', compact('kozijnen', 'project', 'info', 'akkoord'));
+        if ($project){
+            $kozijnen = $project->kozijnen;
+            $akkoord = $project->akkoord;
+            $info = $project->info;
+            return view('projects/show', compact('kozijnen', 'project', 'info', 'akkoord'));
+        }
+        else{
+            return redirect('/')->with('message', 'Project does not exist!');
+        }
     }
 
     public function create(Request $request)
@@ -57,31 +64,57 @@ class ProjectController extends Controller
 
     public function edit(Project $project)
     {
-        $kozijnen1 = Kozijnen::all();
-        $attributes = array();
-        foreach ($kozijnen1 as $koz) {
-            array_push($attributes, Kozijnen::find($koz->id)->attributes);
+        if ($project){
+            $all = Kozijnen::all();
+            $attributes = array();
+            foreach ($all as $koz) {
+                array_push($attributes, Kozijnen::find($koz->id)->attributes);
+            }
+            $kozijnen = $project->kozijnen;
+            return view('projects/edit', compact('kozijnen', 'project', 'attributes'));
         }
-        $kozijnen = $project->kozijnen;
-        return view('projects/edit', compact('kozijnen', 'project', 'attributes'));
+        else{
+            return redirect('/')->with('message', 'Project does not exist!');
+        }
     }
 
     public function update(Request $request, Project $project)
     {
-        $formFields = $request->validate([
-            'KlantNaam' => 'required',
-            'ProjectAdres' => 'required',
-            'Email' => ['required', 'email'],
-            '*' => 'required',
-        ]);
-        $project->update($formFields);
+        if ($project){
+            $formFields = $request->validate([
+                'KlantNaam' => 'required',
+                'ProjectAdres' => 'required',
+                'Email' => ['required', 'email'],
+            ]);
+            
 
-        return redirect('/')->with('message', 'Project updated successfully!');
+
+
+            $attributes = Kozijnen::all();
+            $attributeValues = $request->except(['_token', 'KlantNaam', 'ProjectAdres', 'Email']);
+    
+            foreach ($attributes as $attribute) {
+                $value = $attributeValues[$attribute->kozijn] ?? null;
+                $project->kozijnen()->syncWithoutDetaching([$attribute->id => ['value' => $value]]);
+            }
+        
+            $project->update($formFields);
+
+            return redirect('/')->with('message', 'Project updated successfully!');
+        }
+        else{
+            return redirect('/')->with('message', 'Project does not exist!');
+        }
     }
 
     public function destroy(Project $project)
     {
-        $project->delete();
-        return redirect('/')->with('message', 'Project deleted successfully');
+        if ($project){
+            $project->delete();
+            return redirect('/')->with('message', 'Project deleted successfully');
+        }
+        else{
+            return redirect('/')->with('message', 'Project does not exist!');
+        }
     }
 }
